requirements.txt
python-dotenv>=1.0.0
aiogram>=3.0.0
python-docx>=1.0.0
pydantic>=2.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
python-dateutil>=2.8.2
PyYAML>=6.0.1
aiofiles>=24.1.0